requirements.txt
asgiref==3.8.1
branca==0.7.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cryptography==43.0.0
Django==5.0.6
django-bootstrap-modal-forms==3.0.4
django-crispy-forms==2.1
django-htmx==1.17.3
djangorestframework==3.15.1
folium==0.16.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mercadopago==2.2.3
numpy==1.26.4
pandas==2.2.2
pdfminer==20191125
pdfminer.six==20240706
pdfrw==0.4
pillow==10.3.0
polyline==2.0.2
pycparser==2.22
pycryptodome==3.20.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-form==0.2.3
pytz==2024.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.6.0
